refactor(internal): split the text layer out of blocksignificand and blockfraction (#1334)
Both were the same shape blockbinary had before #1394: a core carrying its own
to_binary/to_hex and an in-class friend operator<<, which put <sstream> into the
graph of everything built on them.

    blocksignificand.hpp   55362 -> 25634 lines  (-54%)   3 -> 0 I/O-family
    blockfraction.hpp      55337 -> 25608 lines  (-54%)   3 -> 0 I/O-family

Four of the five internal cores now meet both acceptance criteria:

    blocksignificand.hpp   25634   0 I/O
    blockfraction.hpp      25608   0 I/O
    blockbinary.hpp        33243   0 I/O
    ieee754_core.hpp       43486   0 I/O
    blocktriple.hpp        98791   4 I/O   <- its own text layer remains

Also in this change:

  blocktriple.hpp now includes native/ieee754_core.hpp rather than the ieee754.hpp
  umbrella. That alone is worth 1141 lines and no I/O reduction, because
  blocktriple's own text is what holds the four; it is a prerequisite, not a
  payoff.

  ieee754_core.hpp gained integer_type_tag.hpp. blocktriple's type_tag calls
  type_tag(bt{}) on its block type, and the umbrella supplied that where the core
  did not. THE GAP EXISTED THE MOMENT #1400 MERGED and the sweep could not see
  it: nothing included the core yet, so "17 failures, identical to baseline" was
  true and uninformative. A header nothing consumes is untested by construction.

  TRACE_DIV and its guarded code are removed from blocksignificand, blockfraction
  and blocktriple, per the same call made for blockbinary in 28687f3b. That also
  removes the need for a conditional <iostream> in the new text headers.

  The in-class friend operator<< is gone from both: each was
  `return ostr << double(v);`, i.e. the public conversion, so friendship was
  never needed -- and it was what let a translation unit missing the include
  compile and then fail to link.

A REAL BUG I INTRODUCED AND CAUGHT. The tail extraction lifted
twosComplementFree into the text layer. It is arithmetic, not text; it merely sat
after the "conversions to string representations" banner, so I selected it by
POSITION rather than by KIND. Two constexpr tests failed with "used before its
definition". Returned to the core in both files, with a comment saying why.

That is the same mistake I caught by hand in the posit split -- ulp() sitting
inside the I/O region -- and missed here because I trusted the banner. For the
rest of the rollout: a region banner is a hint, not a boundary. Classify each
function by what it returns, not by where it sits.

Blast radius handled compile-driven, in two passes: 55 translation units gained
the include the compiler asked for.

Verified: the six headers and all 38 umbrellas compile on gcc 13.3, clang 18.1
and riscv64-linux-gnu cross; the 1057-TU sweep across ten source trees gives 17
failures, IDENTICAL to the main baseline; check-ascii clean.

Relates to #1334
