+++ education/internal/blockfraction.cpp
@@ -9,6 +9,8 @@
 #include <string>
 #include <universal/internal/blockfraction/blockfraction.hpp>
 #include <universal/verification/test_suite.hpp>
+#include <universal/internal/blockfraction/manipulators.hpp>   // (#1334)
+#include <universal/internal/blockfraction/iostream.hpp>   // (#1334)
 
 using namespace sw::universal;
