+++ include/sw/universal/internal/blockfraction/blockfraction.hpp
@@ -7,19 +7,7 @@
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <cstdint>
 #include <cmath> // for std::pow() used in conversions to native IEEE-754 formats values
-// TRACE_DIV: compile-time tracing of the division algorithm, default off.
-// Declared ahead of the includes because <iostream> is needed only when it is
-// on. #ifndef-guarded so a caller can enable it by defining the macro first;
-// the previous unconditional #define silently reset it to 0.
-#ifndef TRACE_DIV
-#define TRACE_DIV 0
-#endif
 
-#if TRACE_DIV
-#include <iostream>
-#endif
-#include <string>
-#include <sstream>
 
 #include <universal/internal/blockfraction/blockfraction_fwd.hpp>
 #include <universal/internal/blocktype/carry.hpp>
@@ -628,117 +616,16 @@ public:
 	}
 
 
-	// ostream operator
-	friend std::ostream& operator<<(std::ostream& ostr, const blockfraction& v) {
-		return ostr << double(v);
-	}
+	// no friend declaration for operator<<: it is defined in
+	// internal/blockfraction/iostream.hpp and needs no private access (it streams the
+	// public double() conversion). Declaring it here would let a translation unit
+	// that forgot that include COMPILE and then fail to link -- a compile error
+	// naming the missing function is the better failure. See #1334.
 };
 
-//////////////////////////////////////////////////////////////////////////////
-// conversions to string representations
-
-// create a binary representation of the blockfraction: 00h.ffff
-// by design, the radix point is at nbits-3
-template<unsigned nbits, typename bt>
-std::string to_binary(const blockfraction<nbits, bt>& number, bool nibbleMarker = false) {
-	std::stringstream s;
-	s << "0b";
-	for (int i = nbits - 1; i >= 0; --i) {
-		s << (number.at(unsigned(i)) ? '1' : '0');
-		if (i == number.radix()) {
-			s << '.';
-		}
-		else {
-			if (i > 0 && (i % 4) == 0 && nibbleMarker) s << '\'';
-		}
-	}
-	return s.str();
-}
-
-// local helper to display the contents of a byte array
-template<unsigned nbits, typename bt>
-std::string to_hex(const blockfraction<nbits, bt>& number, bool wordMarker = true) {
-	static constexpr unsigned bitsInByte = 8;
-	static constexpr unsigned bitsInBlock = sizeof(bt) * bitsInByte;
-	char hexChar[16] = {
-		'0', '1', '2', '3', '4', '5', '6', '7',
-		'8', '9', 'A', 'B', 'C', 'D', 'E', 'F',
-	};
-	std::stringstream ss;
-	ss << "0x" << std::hex;
-	int nrNibbles = int(1 + ((nbits - 1) >> 2));
-	for (int n = nrNibbles - 1; n >= 0; --n) {
-		uint8_t nibble = number.nibble(static_cast<unsigned>(n));
-		ss << hexChar[nibble];
-		if (wordMarker && n > 0 && ((n * 4ll) % bitsInBlock) == 0) ss << '\'';
-	}
-	return ss.str();
-}
-
-///////////////////////////////////////////////////////////////////////////////
-// specialty binary operators
-
-
-// unrounded division, returns a blockfraction that is of size 2*nbits
-template<unsigned nbits, unsigned roundingBits, typename bt>
-blockfraction<2 * nbits + roundingBits, bt> urdiv(const blockfraction<nbits, bt>& a, const blockfraction<nbits, bt>& b, blockfraction<roundingBits, bt>& r) {
-	if (b.iszero()) {
-		// division by zero
-		throw "urdiv divide by zero";
-	}
-	// generate the absolute values to do long division 
-	// 2's complement special case -max requires an signed int that is 1 bit bigger to represent abs()
-	bool a_sign = a.sign();
-	bool b_sign = b.sign();
-	bool result_negative = (a_sign ^ b_sign);
-
-	// normalize both arguments to positive in new size
-	blockfraction<nbits + 1, bt> a_new(a); // TODO optimize: now create a, create _a.bb, copy, destroy _a.bb_copy
-	blockfraction<nbits + 1, bt> b_new(b);
-	if (a_sign) a_new.twoscomplement();
-	if (b_sign) b_new.twoscomplement();
-
-	// initialize the long division
-	blockfraction<2 * nbits + roundingBits, bt> decimator(a_new);
-	blockfraction<2 * nbits + roundingBits, bt> subtractand(b_new); // prepare the subtractand
-	blockfraction<2 * nbits + roundingBits, bt> result;
-
-	int msp = nbits + roundingBits - 1; // msp = most significant position
-	decimator <<= msp; // scale the decimator to the largest possible positive value
-
-	int msb_b = subtractand.msb();
-	int msb_a = decimator.msb();
-	int shift = msb_a - msb_b;
-	int scale = shift - msp;   // scale of the result quotient
-	subtractand <<= shift;
-
-#if TRACE_DIV
-	std::cout << "  " << to_binary(decimator) << std::endl;
-	std::cout << "- " << to_binary(subtractand) << " shift: " << shift << std::endl;
-#endif
-	// long division
-	for (int i = msb_a; i >= 0; --i) {
-
-		if (subtractand <= decimator) {
-			decimator -= subtractand;
-			result.set(static_cast<unsigned>(i));
-		}
-		else {
-			result.reset(static_cast<unsigned>(i));
-		}
-		subtractand >>= 1;
-
-#if TRACE_DIV
-		std::cout << "  " << to_binary(decimator) << ' ' << to_binary(result) << std::endl;
-		std::cout << "- " << to_binary(subtractand) << std::endl;
-#endif
-	}
-	result <<= scale;
-	if (result_negative) result.twosComplement();
-	r.assign(result); // copy the lowest bits which represent the bits on which we need to apply the rounding test
-	return result;
-}
-
+// twosComplementFree: arithmetic, not text. It sat after the "conversions to
+// string representations" banner in the original file, so the text lift caught
+// it by position rather than by kind; it belongs here (#1334).
 // free function generator of the 2's complement of a blockfraction
 template<unsigned nbits, typename bt>
 constexpr blockfraction<nbits, bt> twosComplementFree(const blockfraction<nbits, bt>& a) noexcept {
