+++ internal/blockfraction/conversion/conversion.cpp
@@ -10,6 +10,8 @@
 
 #include <universal/internal/blockfraction/blockfraction.hpp>
 #include <universal/verification/test_suite.hpp>
+#include <universal/internal/blockfraction/iostream.hpp>   // (#1334)
+#include <universal/internal/blockfraction/manipulators.hpp>   // (#1334)
 
 /*
 A blockfraction is an unsigned binary encoding with a radix point
