+++ internal/blocksignificand/api/api.cpp
@@ -11,6 +11,8 @@
 #include <universal/native/integers.hpp>
 #include <universal/internal/blocksignificand/blocksignificand.hpp>
 #include <universal/verification/test_suite.hpp>
+#include <universal/internal/blocksignificand/manipulators.hpp>   // (#1334)
+#include <universal/internal/blocksignificand/iostream.hpp>   // (#1334)
 
 /*
