+++ internal/blocksignificand/arithmetic/division.cpp
@@ -16,6 +16,8 @@
 #include <universal/verification/test_reporters.hpp>
 #include <universal/verification/blocksignificand_test_suite.hpp>
 #include <universal/internal/blockbinary/iostream.hpp>   // operator<< on blockbinary (#1334)
+#include <universal/internal/blocksignificand/iostream.hpp>   // (#1334)
+#include <universal/internal/blocksignificand/manipulators.hpp>   // (#1334)
 
 template<unsigned nbits, typename BlockType>
 void TestMostSignificantBit() {
