+++ internal/blocktriple/arithmetic/division.cpp
@@ -21,6 +21,7 @@
 #include <universal/internal/blocktriple/blocktriple_debug.hpp>  // constexprClassParameters + tracing (#1334)
 #include <universal/verification/test_status.hpp>
 #include <universal/verification/test_reporters.hpp>
+#include <universal/internal/blocksignificand/manipulators.hpp>   // (#1334)
 
 // enumerate all division cases for an blocktriple<nbits,BlockType> configuration
 // TODO: fix test failures in VerifyMultiplication<BlockTripleConfiguration>
