+++ internal/blocktriple/conversion/conversion.cpp
@@ -16,6 +16,7 @@
 #include <universal/native/integers.hpp>
 #include <universal/internal/blocktriple/blocktriple.hpp>
 #include <universal/verification/test_suite.hpp>
+#include <universal/internal/blocksignificand/manipulators.hpp>   // (#1334)
 
 namespace sw { namespace universal {
