+++ static/float/cfloat/conversion/saturating/maxexpvalue/from_blocktriple.cpp
@@ -19,6 +19,7 @@
 #include <universal/verification/test_suite_randoms.hpp>
 #include <universal/verification/cfloat_test_suite.hpp>
 #include <universal/number/cfloat/table.hpp>
+#include <universal/internal/blocksignificand/manipulators.hpp>   // (#1334)
 
 /*
 How do you test the conversion state space of blocktriple to cfloat.
